pyproject.toml
[tool.poetry]
name = "EigenLedger"
authors = ["Santosh <santoshpassoubadyp@gmail.com>"]
version = "2.1.6"
description = "An Open Source Portfolio Management Framework for Everyone 投资组合管理"
readme = "README.md"
license = "LICENSE"
homepage = "https://github.com/ssantoshp/EigenLedger"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.21.0"
matplotlib = "^3.4.0"
quantstats = "^0.0.62"
yfinance = "^0.2.0"
ipython = "^7.16.0"
fpdf = "^1.7.2"
pyportfolioopt = "^1.4.0"
google-api-python-client = "^2.0.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"

[tool.poetry.scripts]
portfolio-tracker = "EigenLedger.portfolio_tracker:main"

[[tool.poetry.packages]]
include = "EigenLedger"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

